% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\docType{data}
\name{dxd}
\alias{dxd}
\title{A DEXSeqDataSet containing pasilla exon-level counts}
\format{
A DEXSeqDataSet
}
\usage{
data(pasillaDEXSeqDataSet)
}
\description{
A DEXSeqDataSet containing exon level counts of the pasilla KO
experiment.
}
\keyword{datasets}
