\name{pasillaDEXSeqDataSet}
\docType{data}
\alias{pasillaDEXSeqDataSet}
\title{Read counts per exon, or per gene, from RNA-seq samples}
\description{
\code{pasillaDEXSeqDataSet} is a \code{DEXSeqDataSet} object with exon counts.
}
\usage{
data("pasillaDEXSeqDataSet")
}
\format{DEXSeqDataSet}
\source{Processed data from NCBI Gene Expression Omnibus under accession numbers GSM461176 to GSM461181.}
\references{Brooks et al. Conservation of an RNA regulatory map between Drosophila and mammals. Genome Research, 2010}
\keyword{datasets}
